requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pyTelegramBotAPI==3.6.2
python-dotenv==0.8.2
requests==2.18.4
six==1.11.0
slackclient==1.2.1
urllib3==1.22
websocket-client==0.47.0
